requirements.txt
# マーケティングエージェント 依存関係（ルート）

PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2

# オプション
# beautifulsoup4==4.12.2
# selenium==4.15.0
# pillow==10.0.1


